refactor(memory): register job handlers directly in the memory plugin's init hook
Move memory job-handler registration and worker startup fully inside the
memory plugin instead of routing them through a generic plugin job-handler
registry. The plugin's `init` hook now registers its own handlers directly
into the worker dispatch table (synchronously, before `runMemoryStartup` is
kicked off), and the jobs worker is started near the end of `runMemoryStartup`
where it belongs — so registration is guaranteed to happen-before the worker's
first claim. Keeping registration inside the plugin means the startup path no
longer reaches back through the `plugins/defaults` barrel, which had closed an
import cycle when startup moved into the init hook.

Since the memory plugin was the only contributor of `jobHandlers`, the generic
plugin job-handler registry became dead once the memory→worker forwarding was
removed (the worker is owned by the memory plugin and starts mid-bootstrap,
before user plugins init, so a shared registry could never forward to it in
time). Remove it: the plugin `jobHandlers` field, `job-handler-registry.ts`,
`registerDefaultPluginJobHandlers`, and the bootstrap wiring. The host's
non-plugin domain handlers are split into `registerDomainJobHandlers`, called
from `lifecycle.ts` before plugins boot; the standalone worker process
self-registers both handler sets plus the persistence hooks.
