+++ assistant/src/plugins/defaults/index.ts
@@ -33,10 +33,6 @@ import {
   clearInjectorRegistry,
   registerPluginInjectors,
 } from "../injector-registry.js";
-import {
-  clearJobHandlerRegistry,
-  registerPluginJobHandlers,
-} from "../job-handler-registry.js";
 import { registerPlugin, resetPluginRegistryForTests } from "../registry.js";
 import { type Plugin, PluginExecutionError } from "../types.js";
 import { channelInjectors } from "./channel/injectors.js";
@@ -74,7 +70,6 @@ import memoryPostCompact from "./memory/hooks/post-compact.js";
 import memoryShutdown from "./memory/hooks/shutdown.js";
 import memoryUserPromptSubmit from "./memory/hooks/user-prompt-submit.js";
 import { memoryInjectors } from "./memory/injectors.js";
-import { memoryJobHandlers } from "./memory/job-handlers.js";
 import memoryPkg from "./memory/package.json" with { type: "json" };
 import { memoryPersistenceHooks } from "./memory/persistence-hooks.js";
 import {
@@ -183,7 +178,6 @@ export const defaultMemoryPlugin: Plugin = {
     "post-compact": memoryPostCompact,
   },
   injectors: [...memoryInjectors, memoryV3Injector, memoryV3SpotlightInjector],
-  jobHandlers: memoryJobHandlers,
 };
 
 /**
@@ -486,26 +480,6 @@ export function registerDefaultPluginInjectors(): void {
 }
 
 /**
- * Register every default plugin's background-job handlers into the global
- * job-handler registry — the job-handler analog of
- * {@link registerDefaultPluginInjectors}. `bootstrapPlugins` calls this before
- * the per-plugin init loop so a default plugin's handlers are present
- * regardless of its disabled-state. The standalone memory-worker process does
- * not run plugin bootstrap, so `registerMemoryJobHandlers`
- * (`jobs/register-job-handlers.ts`) calls this directly before forwarding the
- * registry union into the worker dispatch table. Idempotent:
- * `registerPluginJobHandlers` replaces a plugin's prior set, so the per-plugin
- * re-registration in `initializePlugin` is a harmless no-op replace.
- */
-export function registerDefaultPluginJobHandlers(): void {
-  for (const plugin of getAllDefaultPlugins()) {
-    if (plugin.jobHandlers && plugin.jobHandlers.length > 0) {
-      registerPluginJobHandlers(plugin.manifest.name, plugin.jobHandlers);
-    }
-  }
-}
-
-/**
  * Wrap a plugin's persistence hooks so its ACTIVE side-effect hooks no-op while
  * that plugin is disabled (`assistant plugins disable <name>`), mirroring the
  * read-time disabled-state filtering the injector/hook/job-handler/tool surfaces
@@ -560,7 +534,7 @@ export function guardPersistenceHooksByDisabledState(
 /**
  * Install the memory feature's persistence-lifecycle handlers into the
  * persistence seam — the persistence-hooks analog of
- * {@link registerDefaultPluginJobHandlers}. `bootstrapPlugins` calls this
+ * {@link registerDefaultPluginInjectors}. `bootstrapPlugins` calls this
  * before the per-plugin init loop so the seam is wired up front; the handlers
  * are guarded by {@link guardPersistenceHooksByDisabledState} so a disabled
  * memory plugin drives no persistence side effects and re-enabling it takes
@@ -599,7 +573,5 @@ export function resetPluginRegistryAndRegisterDefaults(): void {
   registerDefaultPlugins();
   clearInjectorRegistry();
   registerDefaultPluginInjectors();
-  clearJobHandlerRegistry();
-  registerDefaultPluginJobHandlers();
   registerDefaultPluginPersistenceHooks();
 }
